pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "channels[daphne]>=4.2.0",
    "django>=5.1.7",
    "django-ninja>=1.3.0",
]
